Added some programming/cyber links; removed the Mac hate.
The Mac stuff makes us sound out of touch in my view. They have been the
most popular systems at the last several places I've worked/inteviewed. If we
don't have the ability to tech support them, that's fine (and we should
definitely be open about that), but it's weird to be so blantantly
"don't get this or you're doomed". I personally recommend them when
prospective students or their parents ask me what type of laptop to get
before starting at a university.